refactor components to get rid of tab wonkiness
